+++ plone/app/event/bbb/portlets/portlet_calendar.py
@@ -1,19 +1,27 @@
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from plone.app.portlets import PloneMessageFactory as _
-from plone.app.portlets.browser.z3cformhelper import AddForm
-from plone.app.portlets.browser.z3cformhelper import EditForm
+try:
+    from plone.app.portlets.browser import z3cformhelper
+    P_A_PORTLETS_PRE_3 = True
+except:
+    P_A_PORTLETS_PRE_3 = False
+from plone.app.portlets.portlets import base
 from plone.app.event.portlets.portlet_calendar import Renderer as RendererBase
 from plone.app.event.portlets.portlet_calendar import ICalendarPortlet
 from plone.app.event.portlets.portlet_calendar import Assignment
-from z3c.form import field
+if P_A_PORTLETS_PRE_3:
+    from z3c.form import field
 
 
 class Renderer(RendererBase):
     render = ViewPageTemplateFile('portlet_calendar.pt')
 
 
-class AddForm(AddForm):
-    fields = field.Fields(ICalendarPortlet)
+class AddForm(base.AddForm):
+    if P_A_PORTLETS_PRE_3:
+        fields = field.Fields(ICalendarPortlet)
+    else:
+        schema = ICalendarPortlet
     label = _(u"Add Calendar Portlet")
     description = _(u"This portlet displays events in a calendar.")
 
@@ -22,7 +30,10 @@ class AddForm(AddForm):
                           search_base_uid=data.get('search_base_uid', None))
 
 
-class EditForm(EditForm):
-    fields = field.Fields(ICalendarPortlet)
+class EditForm(base.EditForm):
+    if P_A_PORTLETS_PRE_3:
+        fields = field.Fields(ICalendarPortlet)
+    else:
+        schema = ICalendarPortlet
     label = _(u"Edit Calendar Portlet")
     description = _(u"This portlet displays events in a calendar.")